Parser.Token

Description

The module Parser.Token contains parsers for tokens.

Documentation

(**>) :: Applicative f => f a -> f b -> f b

**> is a version of *> with increased fixity.

digit2Num :: Num a => Char -> a

pNumeralStr :: Parser String

pNumeralStr ~=> (0|([1-9][0-9]*))

pSNumeral :: Num a => Parser a

pSNumeral converts a string form of a numeral into a Num a.

pSDecimal :: Parser Double

pSDecimal ~=> (0|([1-9][0-9]*))[.]([0-9]+).

pBDigit :: Parser BDigit

pBDigit ~=> [01].

pSBinary :: Parser SBinary

pSBinary ~=> #b[01]+. We consider binary as a string.

pHDigit :: Parser HDigit

pHDigit ~=> [0-9a-fA-F]

pSHex :: Parser SHex

pSHex ~=> #x[0-9a-fA-F]. We consider hex as a string.

pString :: Parser String

pString parses a sequence of pCharString enclosed in ".

pSSymbol :: Parser SSymbol

pSSymbol parses a SSymbol.

pReservSym :: Parser ReservSym

pReservSym parses a reserved word.

pSimpleSym :: Parser SimpleSym

pSymSimple parses a non-empty sequence of symbol characters, not beginning with a digit, and not a reserved word.

pQuotedSym :: Parser QuotedSym

pSymQuoted parses a sequence of smt-lib characters non including \ and |, inclosed in | characters.

pSKeyword :: Parser SKeyword

pSKeyword parses a colon followed by a non-empty sequence of symbol characters.